+++ commons/ava_testsuite/fully_connected_test/fully_connected_test_.go
@@ -50,12 +50,12 @@ func (test StakingNetworkFullyConnectedTest) Run(network networks.Network, conte
 	}
 
 	nonBootValidatorClient := allGeckoClients[nonBootValidatorServiceId]
-	highLevelExtraStakerClient := rpc_workflow_runner.NewRpcWorkflowRunner(
+	highLevelExtraStakerClient := rpc_workflow_runner.NewRPCWorkFlowRunner(
 		nonBootValidatorClient,
 		stakerUsername,
 		stakerPassword,
 		networkAcceptanceTimeout)
-	if err := highLevelExtraStakerClient.GetFundsAndStartValidating(seedAmount, stakeAmount); err != nil {
+	if _, err := highLevelExtraStakerClient.ImportGenesisFundsAndStartValidating(seedAmount, stakeAmount); err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Failed to add extra staker."))
 	}
 
@@ -119,11 +119,11 @@ func getNodeIdsAndClients(
 			testContext.Fatal(stacktrace.Propagate(err, "An error occurred getting the Gecko client for service with ID %v", serviceId))
 		}
 		allGeckoClients[serviceId] = client
-		nodeId, err := client.InfoAPI().GetNodeID()
+		nodeID, err := client.InfoAPI().GetNodeID()
 		if err != nil {
 			testContext.Fatal(stacktrace.Propagate(err, "An error occurred getting the Gecko node ID for service with ID %v", serviceId))
 		}
-		allNodeIds[serviceId] = nodeId
+		allNodeIds[serviceId] = nodeID
 	}
 	return
 }